Move methods for actions to base class for Upstream/Distgit
