+++ packit/base_git.py
@@ -22,12 +22,13 @@
 
 import git
 import logging
-from typing import Optional
+from typing import Optional, Callable
 
 from rebasehelper.specfile import SpecFile
 from packit.config import Config, PackageConfig
 from packit.exceptions import PackitException
 from packit.local_project import LocalProject
+from packit.utils import run_command
 
 logger = logging.getLogger(__name__)
 
@@ -118,3 +119,80 @@ class PackitRepositoryBase:
         #       distribute it to the container, prepare git config and then we can start signing
         # TODO: make -s configurable
         self.local_project.git_repo.git.commit(*commit_args)
+
+    def run_action(self, action_name: str, method: Callable = None, *args, **kwargs):
+        """
+        Run the method in the self._with_action block.
+
+        Usage:
+
+        >   self._run_action(
+        >        action_name="sync", method=dg.sync_files, upstream_project=up.local_project
+        >   )
+        >   # If user provided custom command for the `sync`, it will be used.
+        >   # Otherwise, the method `dg.sync_files` will be used
+        >   # with parameter `upstream_project=up.local_project`
+        >
+        >   self._run_action(action_name="pre-sync")
+        >   # This will be used as an optional hook
+
+        :param action_name: action_name: str (Name of the action that can be overwritten
+                                                in the package_config.actions)
+        :param method: method to run if the action was not defined by user
+                    (if not specified, the action can be used for custom hooks)
+        :param args: args for the method
+        :param kwargs: kwargs for the method
+        """
+        if not method:
+            logger.debug(f"Running {action_name} hook.")
+        if self.with_action(action_name=action_name):
+            if method:
+                method(*args, **kwargs)
+
+    def has_action(self, action_name: str) -> bool:
+        """
+        Is the action defined in the config?
+        """
+        return action_name in self.package_config.actions
+
+    def with_action(self, action_name: str) -> bool:
+        """
+        If the action is defined in the self.package_config.actions,
+        we run it and return False (so we can skip the if block)
+
+        If the action is not defined, return True.
+
+        Usage:
+
+        >   if self._with_action(action_name="patch"):
+        >       # Run default implementation
+        >
+        >   # Custom command was run if defined in the config
+
+        Context manager is currently not possible without ugly hacks:
+        https://stackoverflow.com/questions/12594148/skipping-execution-of-with-block
+        https://www.python.org/dev/peps/pep-0377/ (rejected)
+
+        :param action_name: str (Name of the action that can be overwritten
+                                                in the package_config.actions)
+        :return: True, if the action is not overwritten, False when custom command was run
+        """
+        logger.debug(f"Running {action_name}.")
+        if action_name in self.package_config.actions:
+            command = self.package_config.actions[action_name]
+            logger.info(f"Using user-defined script for {action_name}: {command}")
+            run_command(cmd=command)
+            return False
+        logger.debug(f"Running default implementation for {action_name}.")
+        return True
+
+    def get_output_from_action(self, action_name: str):
+        """
+        Run action if specified in the self.actions and return output
+        else return None
+        """
+        if action_name in self.package_config.actions:
+            command = self.package_config.actions[action_name]
+            logger.info(f"Using user-defined script for {action_name}: {command}")
+            return run_command(cmd=command, output=True)
+        return None
