+++ packit/config.py
@@ -26,8 +26,7 @@ import os
 from enum import IntEnum
 from functools import lru_cache
 from pathlib import Path
-
-from typing import Optional, List, NamedTuple, Dict, Callable
+from typing import Optional, List, NamedTuple, Dict
 
 import click
 import jsonschema
@@ -37,7 +36,7 @@ from yaml import safe_load
 from ogr.abstract import GitProject
 from packit.constants import CONFIG_FILE_NAMES
 from packit.exceptions import PackitConfigException, PackitException
-from packit.utils import exclude_from_dict, run_command
+from packit.utils import exclude_from_dict
 
 logger = logging.getLogger(__name__)
 
@@ -371,83 +370,6 @@ class PackageConfig:
     @classmethod
     def validate_dict(cls, raw_dict: dict) -> None:
         jsonschema.validate(raw_dict, PACKAGE_CONFIG_SCHEMA)
-
-    def run_action(self, action_name: str, method: Callable = None, *args, **kwargs):
-        """
-        Run the method in the self._with_action block.
-
-        Usage:
-
-        >   self._run_action(
-        >        action_name="sync", method=dg.sync_files, upstream_project=up.local_project
-        >   )
-        >   # If user provided custom command for the `sync`, it will be used.
-        >   # Otherwise, the method `dg.sync_files` will be used
-        >   # with parameter `upstream_project=up.local_project`
-        >
-        >   self._run_action(action_name="pre-sync")
-        >   # This will be used as an optional hook
-
-        :param action_name: action_name: str (Name of the action that can be overwritten
-                                                in the package_config.actions)
-        :param method: method to run if the action was not defined by user
-                    (if not specified, the action can be used for custom hooks)
-        :param args: args for the method
-        :param kwargs: kwargs for the method
-        """
-        if not method:
-            logger.debug(f"Running {action_name} hook.")
-        if self.with_action(action_name=action_name):
-            if method:
-                method(*args, **kwargs)
-
-    def has_action(self, action_name: str) -> bool:
-        """
-        Is the action defined in the config?
-        """
-        return action_name in self.actions
-
-    def with_action(self, action_name: str) -> bool:
-        """
-        If the action is defined in the self.package_config.actions,
-        we run it and return False (so we can skip the if block)
-
-        If the action is not defined, return True.
-
-        Usage:
-
-        >   if self._with_action(action_name="patch"):
-        >       # Run default implementation
-        >
-        >   # Custom command was run if defined in the config
-
-        Context manager is currently not possible without ugly hacks:
-        https://stackoverflow.com/questions/12594148/skipping-execution-of-with-block
-        https://www.python.org/dev/peps/pep-0377/ (rejected)
-
-        :param action_name: str (Name of the action that can be overwritten
-                                                in the package_config.actions)
-        :return: True, if the action is not overwritten, False when custom command was run
-        """
-        logger.debug(f"Running {action_name}.")
-        if action_name in self.actions:
-            command = self.actions[action_name]
-            logger.info(f"Using user-defined script for {action_name}: {command}")
-            run_command(cmd=command)
-            return False
-        logger.debug(f"Running default implementation for {action_name}.")
-        return True
-
-    def get_output_from_action(self, action_name: str):
-        """
-        Run action if specified in the self.actions and return output
-        else return None
-        """
-        if action_name in self.actions:
-            command = self.actions[action_name]
-            logger.info(f"Using user-defined script for {action_name}: {command}")
-            return run_command(cmd=command, output=True)
-        return None
 
 
 def get_local_package_config(
